refactor(domain): change package of GameResult.kt as domain
