+++ src/main/kotlin/lotto/domain/GameResult.kt
@@ -1,7 +1,5 @@
-package lotto.services
+package lotto.domain
 
-import lotto.domain.Rank
-import lotto.domain.WinningCombination
 import lotto.dto.IssuedTickets
 
 data class GameResult(
