+++ src/test/kotlin/lotto/services/TicketIssuerTest.kt
@@ -1,8 +1,5 @@
 package lotto.services
 
-import lotto.domain.LottoTicket
-import lotto.dto.IssuedTickets
-import org.assertj.core.api.Assertions.assertThat
 import org.junit.jupiter.api.Test
 import org.junit.jupiter.api.assertDoesNotThrow
 import org.junit.jupiter.api.assertThrows
@@ -12,10 +9,10 @@ class TicketIssuerTest {
     fun `throw if can not issue ticket`() {
         assertThrows<IllegalArgumentException> {
             val requests = listOf(
-                    setOf(0, 1, 2, 3, 4, 5, 6),
-                    setOf(46, 45, 44, 43, 42, 41)
-                )
-            val issuedTickets = TicketIssuer.with(requests)
+                setOf(0, 1, 2, 3, 4, 5, 6),
+                setOf(46, 45, 44, 43, 42, 41)
+            )
+            TicketIssuer.with(requests)
         }
     }
 
@@ -23,10 +20,10 @@ class TicketIssuerTest {
     fun `dose not throw if issue ticket`() {
         assertDoesNotThrow {
             val requests = listOf(
-                    setOf(1, 2, 3, 4, 5, 6),
-                    setOf(45, 44, 43, 42, 41, 40)
-                )
-            val issuedTickets = TicketIssuer.with(requests)
+                setOf(1, 2, 3, 4, 5, 6),
+                setOf(45, 44, 43, 42, 41, 40)
+            )
+            TicketIssuer.with(requests)
         }
     }
 }
